hide boost warnings (no more warnings!), update alert key, add generatekey/createalert/signalert/sendalert admin-only commands + more - add root (find root username) admin-only command - fix root user detection - add rootuser to getinfo output - update nsis logos/icons - fix nsis script file - fix genbuild.sh - rearrange code

+++ src/qt/walletmodel.h
@@ -140,7 +140,7 @@ private:
     quint64 cachedBalance;
     quint64 cachedUnconfirmedBalance;
     quint64 cachedImmatureBalance;
-    quint64 cachedNumTransactions;
+    int cachedNumTransactions;
     EncryptionStatus cachedEncryptionStatus;
     int cachedNumBlocks;
